resolved merge : EnnemySawn + EnnemyDestroy system

+++ Dual-Punk/Assets/Scripts/Entities/Player/PlayerState.cs
@@ -47,7 +47,7 @@ public class PlayerState : MonoBehaviour
         if (amount < 0){
             throw new ArgumentOutOfRangeException("Cannot have Negative damage.");
         }
-        this.Health -= amount;
+        this.Health -= (int)amount;
 
         if (Health <= 0){
             Die();
